// src/app/models/veiculo.ts
export class Veiculo {
  placa: string//
  marca: string//
  chassi: string//
  renavam: string//
  cor: string//
  anofabricacao: string//
  anomodelo: string//
  combustivel: string//
  especie: string//
  nomeproprietario: string//
  cidade: string //
  uf: string//
  base: string//
  restricaoadministrativa: string//
  restricaofinanceira: string//
  restricaojuridica: string;//
  ultimoipva: string//
  isentoipva: string//
  totalipva: string//
  totaldpvat: string
  totalicenca: string
  valormulta: string //
  dpvatanterior: string//
  ipvaanterior: string//
  licencaanterior: string//
  multatramitacao: string//
  motor: string//
  // new ones
  roubo: string;
  clone : string;
  categoria: string;
  municipio: string;
  dpvatatual:string;
  restricao1:string;
   tipo: string;
  licencaatual: string
  telefonePropietario:string;//
  dataRouboFurto: string;//
  dataDevolucao:string;//
  
  
  
  

  // constructor() {
  //   this.placa = '';
  //   this.chassi = '';
  //   this.clone = '';
  //   this.roubo = '';
  //   this.marca = '';
  //   this.renavam = '';
  //   this.cor = '';
  //   this.anomodelo = '';
  //   this.anofabricacao = '';
  //   this.categoria = '';
  //   this.combustivel = '';
  //   this.especie = '';
  //   this.tipo = '';
  //   this.nomepropriet = '';
  //   this.restricao1 = '';
  //   this.anoipva = '';
  //   this.anolicenca = '';
  //   this.ultimoipva = '';
  //   this.cotaunica = '';
  //   this.isentoipva = '';
  //   this.restricaoadministrativa = '';
  //   this.restricaofinanceira = '';
  //   this.restricaojuridica = '';
  //   this.totalipva = '';
  //   this.dpvatanterior = '';
  //   this.dpvatual = '';
  //   this.ipvaanterior = '';
  //   this.licencaanterior = '';
  //   this.licencaatual = '';
  //   this.valormulta = '';
  //   this.multatramitacao = '';
  //   this.sbase = '';
  //   this.motor = '';
  //   this.cidade = '';
  //   this.uf = '';
  //   this.data_insercao = '';
  // }

}